+++ src/controls/date.ts
@@ -17,7 +17,7 @@ const tagName = "wup-date";
 declare global {
   namespace WUP.Date {
     interface EventMap extends WUP.BaseCombo.EventMap {}
-    interface ValidityMap extends WUP.BaseCombo.ValidityMap, Pick<WUP.Text.ValidityMap, "_mask" | "_parse"> {
+    interface ValidityMap extends WUP.BaseCombo.ValidityMap {
       /** Enabled if option [min] is pointed; If $value < pointed shows message 'Min date is {x}` */
       min: Date;
       /** Enabled if option [min] is pointed; if $value > pointed shows message 'Max date is {x}` */